Dry run for clean

+++ src/cli/shared/clean.ts
@@ -5,7 +5,11 @@ import { PkgdropOptions } from '../../lib/';
 export async function cleanPackagePath(options: PkgdropOptions) {
   const p = filesystem.path(options.package_path);
   if (p) {
-    print.info(`Cleaning output directory`);
-    await filesystem.remove(p);
+    if (options.dry) {
+      print.info(`Cleaning output directory [dry run]`);
+    } else {
+      print.info(`Cleaning output directory`);
+      await filesystem.remove(p);      
+    }
   }
 }
